requirements.txt
requests
datetime 
python-dotenv